refactoring Sample
