+++ src/main/java/org/earthtime/UPb_Redux/dialogs/aliquotManagers/AliquotLegacyEditorForIDTIMS.java
@@ -33,8 +33,8 @@ import javax.swing.JComponent;
 import javax.swing.JLabel;
 import javax.swing.JOptionPane;
 import javax.swing.JTextField;
-import org.earthtime.UPb_Redux.ReduxConstants;
 import org.earthtime.ETReduxFrame;
+import org.earthtime.UPb_Redux.ReduxConstants;
 import org.earthtime.UPb_Redux.aliquots.Aliquot;
 import org.earthtime.UPb_Redux.aliquots.UPbReduxAliquot;
 import org.earthtime.UPb_Redux.exceptions.BadLabDataException;
@@ -1118,7 +1118,7 @@ public class AliquotLegacyEditorForIDTIMS extends AliquotEditorDialog {
             if ( proceed ) {
                 saveAliquot();
                 saveAliquotFraction( fraction );
-                getSample().editUPbFraction( fraction, 8 );
+                parent.editFraction( fraction, 8 );
                 updateFractionRow(
                         fraction,
                         getMyAliquot().getAliquotFractions().indexOf( fraction ) );
