+++ src/main/java/org/earthtime/UPb_Redux/dialogs/sampleManagers/analysisManagers/SampleAnalysisWorkflowManagerLAICPMS.java
@@ -2050,7 +2050,7 @@ public class SampleAnalysisWorkflowManagerLAICPMS extends DialogEditor implement
             String importFolder = null;
             try {
                 importFolder = //
-                        getMySample().importUPbFractionXMLDataFiles(
+getMySample().importFractionXMLDataFiles(
                                 getImportedXMLFractionsFolder(), aliquotNumber, true);
             } catch (FileNotFoundException fileNotFoundException) {
             } catch (BadLabDataException ex) {
@@ -2197,7 +2197,7 @@ public class SampleAnalysisWorkflowManagerLAICPMS extends DialogEditor implement
                 sampleFile = new File(mySample.getReduxSampleFilePath());
             } else {
                 try {
-                    sampleFile = mySample.saveSampleFileAs();
+                    sampleFile = mySample.saveSampleFileAs(((ETReduxFrame)parentFrame).getMyState().getMRUSampleFolderPath());
                     setSampleFolder(new File(sampleFile.getParent()));
                 } catch (BadLabDataException ex) {
                     new ETWarningDialog(ex).setVisible(true);
@@ -2236,7 +2236,7 @@ public class SampleAnalysisWorkflowManagerLAICPMS extends DialogEditor implement
                 sampleFile = new File(mySample.getReduxSampleFilePath());
             } else {
                 try {
-                    sampleFile = mySample.saveSampleFileAs();
+                    sampleFile = mySample.saveSampleFileAs(((ETReduxFrame)parentFrame).getMyState().getMRUSampleFolderPath());
                     try {
                         setSampleFolder(new File(sampleFile.getParent()));
                     } catch (Exception e) {
