+++ src/main/java/org/earthtime/samples/SampleInterface.java
@@ -49,6 +49,130 @@ public interface SampleInterface {
     public abstract void setSampleName(String sampleName);
 
     /**
+     * sets the <code>sampleType</code> of this <code>Sample</code> to the
+     * argument <code>sampleType</code>
+     *
+     * @pre argument <code>sampleType</code> is a valid <code>sampleType</code>
+     * @post this <code>Sample</code>'s <code>sampleType</code> is set to
+     * argument <code>sampleType</code>
+     * @param sampleType value to which <code>sampleType</code> of this
+     * <code>Sample</code> will be set
+     */
+    public abstract void setSampleType(String sampleType);
+
+    /**
+     * sets the <code>analyzed</code> field of this <code>Sample</code> to the
+     * argument <code>analyzed</code>
+     *
+     * @pre argument <code>analyzed</code> is a valid <code>boolean</code>
+     * @post this <code>Sample</code>'s <code>analyzed</code> field is set to
+     * argument <code>analyzed</code>
+     * @param analyzed value to which <code>analyzed</code> field of this
+     * <code>Sample</code> will be set
+     */
+    public abstract void setAnalyzed(boolean analyzed);
+
+    /**
+     *
+     * @return
+     */
+    public abstract boolean isAnalysisTypeTripolized();
+
+    /**
+     * gets the <code>analyzed</code> field of this <code>Sample</code>
+     *
+     * @pre this <code>Sample</code> exists
+     * @post returns the <code>analyzed</code> field of this <code>Sample</code>
+     * @return  <code>boolean</code> - <code>analyzed</code> field of this
+     * <code>Sample</code>
+     */
+    public abstract boolean isAnalyzed();
+
+    // Aliquots
+    /**
+     * finds the <code>Aliquot</code> named <code>name</code> in the array
+     * <code>aliquots</code>.
+     *
+     * @pre an <code>Aliquot</code> exists in <code>aliquots</code> named
+     * <code>name</code>
+     * @post the <code>Aliquot</code> whose name corresponds to argument
+     * <code>name</code> is found and returned
+     * @param name name of the <code>Aliquot</code> to retrieve
+     * @return  <code>Aliquot</code> - the <code>Aliquot</code> from
+     * <code>aliquots</code> whose name correspongs to the argument
+     * <code>name</code>
+     */
+    public abstract Aliquot getAliquotByName(String name);
+
+    public abstract Vector<Aliquot> getActiveAliquots();
+
+    /**
+     * gets the <code>aliquots</code> of this <code>Sample</code>.
+     *
+     * @pre this <code>Sample</code> exists
+     * @post returns the <code>aliquots</code> of this <code>Sample</code>
+     * @return  <code>Vector</code> - set of <code>Aliquots</code> of this
+     * <code>Sample</code>
+     */
+    public abstract Vector<Aliquot> getAliquots();
+
+    /**
+     * finds the <code>Aliquot</code> numbered <code>aliquotNum</code> in the
+     * array <code>aliquots</code>.
+     *
+     * @pre an <code>Aliquot</code> exists in <code>aliquots</code> numbered
+     * with <code>aliquotNum</code>
+     * @post the <code>Aliquot</code> whose number corresponds to argument
+     * <code>aliquotNum</code> is found and returned
+     * @param aliquotNum number of the <code>Aliquot</code> to retrieve
+     * @return  <code>Aliquot</code> - the <code>Aliquot</code> from
+     * <code>aliquots</code> whose number corresponds to the argument
+     * <code>aliquotNum</code>
+     */
+    public abstract Aliquot getAliquotByNumber(int aliquotNum);
+
+    public abstract String getNameOfAliquotFromSample(int aliquotNum);
+
+    // Fractions
+    /**
+     * adds a <code>UPbFraction</code> to the <code>Sample</code>'s set of
+     * <code>Fractions</code>.
+     *
+     * @pre this <code>Sample</code> exists
+     * @post a new default <code>Fraction</code> is added to this
+     * <code>Sample</code>'s <code>Fractions</code>
+     * @param newFraction the <code>Fraction</code> to add to this
+     * <code>Sample</code>
+     */
+    public abstract void addUPbFraction(Fraction newFraction);
+
+    /**
+     * gets the <code>UPbFractions</code> of this <code>Sample</code>.
+     *
+     * @pre this <code>Sample</code> exists
+     * @post returns the <code>UPbFractions</code> of this <code>Sample</code>
+     * @return  <code>Vector</code> - set of <code>Fractions</code> that make up
+     * the <code>UPbFractions</code> of this <code>Sample</code>
+     */
+    public abstract Vector<Fraction> getUPbFractions();
+
+    // Report Settings
+    /**
+     *
+     * @return
+     */
+    public abstract ReportSettings getReportSettingsModelUpdatedToLatestVersion();
+
+    public abstract void setLegacyStatusForReportTable();
+
+    /**
+     *
+     * @return
+     */
+    public abstract ReportSettings getReportSettingsModel();
+
+    // Archiving
+    /**
      * sets the <code>changed</code> field of each <code>UPbFraction</code> in
      * this <code>Sample</code> to <code>false</code> and saves this
      * <code>Sample</code> as a .redux file to <code>reduxSampleFilePath</code>.
@@ -73,99 +197,6 @@ public interface SampleInterface {
     public abstract String saveTheSampleAsSerializedReduxFile(File file);
 
     /**
-     *
-     * @return
-     */
-    public abstract boolean isAnalysisTypeTripolized();
-
-    /**
-     * sets the <code>sampleType</code> of this <code>Sample</code> to the
-     * argument <code>sampleType</code>
-     *
-     * @pre argument <code>sampleType</code> is a valid <code>sampleType</code>
-     * @post this <code>Sample</code>'s <code>sampleType</code> is set to
-     * argument <code>sampleType</code>
-     * @param sampleType value to which <code>sampleType</code> of this
-     * <code>Sample</code> will be set
-     */
-    public abstract void setSampleType(String sampleType);
-
-    /**
-     * sets the <code>analyzed</code> field of this <code>Sample</code> to the
-     * argument <code>analyzed</code>
-     *
-     * @pre argument <code>analyzed</code> is a valid <code>boolean</code>
-     * @post this <code>Sample</code>'s <code>analyzed</code> field is set to
-     * argument <code>analyzed</code>
-     * @param analyzed value to which <code>analyzed</code> field of this
-     * <code>Sample</code> will be set
-     */
-    public abstract void setAnalyzed(boolean analyzed);
-
-    public abstract void setLegacyStatusForReportTable();
-
-    /**
-     * adds a <code>UPbFraction</code> to the <code>Sample</code>'s set of
-     * <code>Fractions</code>.
-     *
-     * @pre this <code>Sample</code> exists
-     * @post a new default <code>Fraction</code> is added to this
-     * <code>Sample</code>'s <code>Fractions</code>
-     * @param newFraction the <code>Fraction</code> to add to this
-     * <code>Sample</code>
-     */
-    public abstract void addUPbFraction(Fraction newFraction);
-
-    /**
-     * finds the <code>Aliquot</code> named <code>name</code> in the array
-     * <code>aliquots</code>.
-     *
-     * @pre an <code>Aliquot</code> exists in <code>aliquots</code> named
-     * <code>name</code>
-     * @post the <code>Aliquot</code> whose name corresponds to argument
-     * <code>name</code> is found and returned
-     * @param name name of the <code>Aliquot</code> to retrieve
-     * @return  <code>Aliquot</code> - the <code>Aliquot</code> from
-     * <code>aliquots</code> whose name correspongs to the argument
-     * <code>name</code>
-     */
-    public abstract Aliquot getAliquotByName(String name);
-
-    /**
-     *
-     * @return
-     */
-    public abstract ReportSettings getReportSettingsModelUpdatedToLatestVersion();
-
-    public abstract Vector<Aliquot> getActiveAliquots();
-
-    /**
-     *
-     * @return
-     */
-    public abstract ReportSettings getReportSettingsModel();
-
-    /**
-     * gets the <code>analyzed</code> field of this <code>Sample</code>
-     *
-     * @pre this <code>Sample</code> exists
-     * @post returns the <code>analyzed</code> field of this <code>Sample</code>
-     * @return  <code>boolean</code> - <code>analyzed</code> field of this
-     * <code>Sample</code>
-     */
-    public abstract boolean isAnalyzed();
-
-    /**
-     * gets the <code>aliquots</code> of this <code>Sample</code>.
-     *
-     * @pre this <code>Sample</code> exists
-     * @post returns the <code>aliquots</code> of this <code>Sample</code>
-     * @return  <code>Vector</code> - set of <code>Aliquots</code> of this
-     * <code>Sample</code>
-     */
-    public abstract Vector<Aliquot> getAliquots();
-
-    /**
      * sets the <code>sampleIGSN</code> of this <code>Sample</code> to the
      * argument <code>sampleIGSN</code>
      *
@@ -187,6 +218,7 @@ public interface SampleInterface {
      */
     public abstract String getSampleIGSN();
 
+    // Parameter Models
     /**
      * sets the <code>myReduxLabData</code> field of this <code>Sample</code> to
      * the argument <code>myReduxLabData</code>
@@ -199,29 +231,6 @@ public interface SampleInterface {
      * this <code>Sample</code> will be set
      */
     public abstract void setMyReduxLabData(ReduxLabData myReduxLabData);
-
-    /**
-     * gets the <code>sampleAgeInterpretationGUISettings</code> of this
-     * <code>Sample</code>
-     *
-     * @pre this <code>Sample</code> exists
-     * @post returns the <code>sampleAgeInterpretationGUISettings</code> of this
-     * <code>Sample</code>
-     * @return  <code>SampleDateInterpretationGUIOptions</code> -
-     * <code>sampleAgeInterpretationGUIOptions</code> of this
-     * <code>Sample</code>
-     */
-    public abstract SampleDateInterpretationGUIOptions getSampleDateInterpretationGUISettings();
-
-    /**
-     * gets the <code>UPbFractions</code> of this <code>Sample</code>.
-     *
-     * @pre this <code>Sample</code> exists
-     * @post returns the <code>UPbFractions</code> of this <code>Sample</code>
-     * @return  <code>Vector</code> - set of <code>Fractions</code> that make up
-     * the <code>UPbFractions</code> of this <code>Sample</code>
-     */
-    public abstract Vector<Fraction> getUPbFractions();
 
     /**
      * gets the <code>physicalConstantsModel</code> of this <code>Sample</code>.
@@ -246,6 +255,7 @@ public interface SampleInterface {
      */
     public abstract ReduxLabData getMyReduxLabData();
 
+    // Graphics
     /**
      *
      * @return
@@ -255,20 +265,16 @@ public interface SampleInterface {
     public abstract GraphAxesSetup getConcordiaGraphAxesSetup();
 
     /**
-     * finds the <code>Aliquot</code> numbered <code>aliquotNum</code> in the
-     * array <code>aliquots</code>.
+     * gets the <code>sampleAgeInterpretationGUISettings</code> of this
+     * <code>Sample</code>
      *
-     * @pre an <code>Aliquot</code> exists in <code>aliquots</code> numbered
-     * with <code>aliquotNum</code>
-     * @post the <code>Aliquot</code> whose number corresponds to argument
-     * <code>aliquotNum</code> is found and returned
-     * @param aliquotNum number of the <code>Aliquot</code> to retrieve
-     * @return  <code>Aliquot</code> - the <code>Aliquot</code> from
-     * <code>aliquots</code> whose number corresponds to the argument
-     * <code>aliquotNum</code>
+     * @pre this <code>Sample</code> exists
+     * @post returns the <code>sampleAgeInterpretationGUISettings</code> of this
+     * <code>Sample</code>
+     * @return  <code>SampleDateInterpretationGUIOptions</code> -
+     * <code>sampleAgeInterpretationGUIOptions</code> of this
+     * <code>Sample</code>
      */
-    public abstract Aliquot getAliquotByNumber(int aliquotNum);
-
-    public abstract String getNameOfAliquotFromSample(int aliquotNum);
+    public abstract SampleDateInterpretationGUIOptions getSampleDateInterpretationGUISettings();
 
 }
